+++ server/api/overzicht/producten/[ean].delete.ts
