+++ server/api/overzicht/producten/[ean].get.ts
