+++ server/api/overzicht/producten/[ean].put.ts
